[FIX/#98] 홈뷰 판매 기능 온보딩뷰로 이관
